+++ client/src/main/scripts/common/video.ts
